ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.8.2
requests==2.32.3
httpx